backend/requirements.txt
fastapi==0.115.2
uvicorn==0.34.3
requests==2.32.3
python-dotenv==1.0.1